WIP: SpeciesNameParser skeleton (not implemented) + test_schema_species_name_parser skeleton (not implemented)
