+++ biosys/apps/main/tests/test_schema_species_name_parser.py
@@ -1,0 +1,11 @@
+from django.test import TestCase
+
+from main.tests.api.helpers import *
+from main.tests.test_data_package import *
+from main.utils_data_package import *
+
+
+class TestSingleSpeciesNameColumn(TestCase):
+
+    def test_happy_path(self):
+        self.fail('Not implemented')
